go/roothash/api: Move runtime messages to its own package
