+++ go/roothash/api/message/message_test.go
@@ -1,4 +1,4 @@
-package block
+package message
 
 import (
 	"testing"
